コントローラの修正と重複部分のリファクタリング#18
