--- back/app/controllers/api/v1/user_bookmarks_controller.rb
@@ -1,8 +1,0 @@
-class Api::V1::UserBookmarksController < ApplicationController
-  before_action :authenticate_api_v1_user!
-
-  def index
-    bookmarks = current_api_v1_user.post_bookmarks.map { |bookmark| bookmark.post_id }
-    render json: { bookmarks: bookmarks }
-  end
-end
