+++ front/src/pages/Home.tsx
@@ -12,7 +12,7 @@ const Home = () => {
     useEffect(() => {
         const fetchBookmarks = async () => {
             try {
-                const response = await api.get('/user_bookmarks');
+                const response = await api.get('/post_bookmarks');
                 const userBookmarks = response.data.bookmarks.reduce((acc: ({ [key: number]: boolean }), postId: number) => {
                     acc[postId] = true;
                     return acc;
